Made a compute shader and a view extension that handles all the color generation
This should be a lot more efficient than the old temp solution I had that used a scene capture. Still need to work on the readback from the GPU, but it works for now at least. With this change, I also added support for opacity, so effects like trails can easily be made.

Also made a manual remap of Razer's mousepads, so they are more 2D. Not sure if that will work on all versions, or just the one that is in the emulator...