Hopefully fixed a Synapse crash. Restructured controller initialization
I couldn't repro the Synapse crash, and Razer's SDK is really inconsistent with how they want you to initialize it, so I just do it in all the ways now... I had hoped to avoid using the ChromaAppInfo.xml, but now I added it just in case. It is marked as a runtime dependency, so it should automatically be staged correctly, I think. Again, I can't actually reproduce these issues, so kind of just trying to fix things blindly...